voice-minimal/requirements.txt
# Minimal dependencies for working voice AI
# Tested with Python 3.10+

# Core framework
pipecat-ai==0.0.85

# LLM providers
google-generativeai>=0.3.0

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
websockets>=12.0

# Environment
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# Telnyx
telnyx>=2.0.0
